# api/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Bikesharing REST API
  version: 0.0.1
tags:
  - name: api
    description: semestral project bikesharing api
paths:
  /api/stands:
    get:
      tags:
        - stands
      summary: Retrieves all stands currently in the system.
      operationId: retrieveStands
      responses:
        '200':
          description: Stands are retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StandDTO'
        '405':
          description: Not supported HTTP method.
        '406':
          description: Not acceptable response representation.
        '500':
          description: Server error.
  /api/bikes:
    get:
      tags:
        - bikes
      summary: Retrieves all rideable bikes currently in the system.
      operationId: retrieveBikes
      responses:
        '200':
          description: Bikes are retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BikeDTO'
        '405':
          description: Not supported HTTP method.
        '406':
          description: Not acceptable response representation.
        '500':
          description: Server error.
  /api/bikes/{standId}:
    get:
      tags:
        - bikes
      summary: Retrieves rideable bikes that belong to stand with given id.
      operationId: retrieveStandBikes
      parameters:
        - name: standId
          in: path
          description: id of stand
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Stand bikes are retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BikeDTO'
        '405':
          description: Not supported HTTP method.
        '406':
          description: Not acceptable response representation.
        '500':
          description: Server error.
  /api/repair:
    get:
      tags:
        - repair
      summary: Retrieves bikes that are due for a repair
      operationId: retrieveBikesForReapir
      responses:
        '200':
          description: Bikes for repair are retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BikeDTO'
        '405':
          description: Not supported HTTP method.
        '406':
          description: Not acceptable response representation.
        '500':
          description: Server error.
  /api/repair/{bikeId}:
    post:
      tags:
        - repair
      summary: Reapirs specific bike with given bike ID.
      operationId: repairBike
      parameters:
        - name: bikeId
          in: path
          description: id of bike
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Bike is repaired.
          content:
            application/json:
              schema:
                type: string
        '405':
          description: Not supported HTTP method.
        '406':
          description: Not acceptable response representation.
        '500':
          description: Server error.
  /api/history/{userId}:
    get:
      tags:
        - history
      summary: Retrieves history of rides ridden by user with given ID.
      operationId: userRideHistory
      parameters:
        - name: userId
          in: path
          description: id of user
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: History of rides is retireved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RideDTO'
        '405':
          description: Not supported HTTP method.
        '406':
          description: Not acceptable response representation.
        '500':
          description: Server error.
components:
  schemas:
    BikeDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        location:
          $ref: '#/components/schemas/LocationDTO'
        lastServiceStamp:
          type: string
          format: date
        stand:
          $ref: '#/components/schemas/StandDTO'
      required:
        - id
        - location
        - lastServiceTimestamp
    StandDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        location:
          $ref: '#/components/schemas/LocationDTO'
      required:
        - id
        - name
        - location
    LocationDTO:
      type: object
      properties:
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
      required:
        - longitude
        - latitude
    RideDTO:
      type: object
      properties:
        bike:
          $ref: '#/components/schemas/BikeDTO'
        startStand:
          $ref: '#/components/schemas/StandDTO'
        endStand:
          $ref: '#/components/schemas/StandDTO'
        startTimeStamp:
          type: string
          format: date-time
        endTimeStamp:
          type: string
          format: date-time
        state:
          type: string
      required:
        - bike
        - startStand
        - endStand
        - startTimeStamp
        - endTimeStamp
        - state